Moved signal-client.js and web-socket.js to shared
