+++ examples/shared/signal-client.js
