+++ examples/shared/web-socket.js
